requirements.txt
discord.py
yt-dlp
static-ffmpeg
asyncio